+++ src/routes/RequestsRoute.js
@@ -44,7 +44,7 @@ import RequestFormContainer from '../RequestFormContainer';
 
 import {
   reportHeaders,
-  fulfilmentTypes,
+  fulfillmentTypes,
   expiredHoldsReportHeaders,
   pickSlipType,
   createModes,
@@ -54,6 +54,7 @@ import {
   DEFAULT_DISPLAYED_YEARS_AMOUNT,
   MAX_RECORDS,
   OPEN_REQUESTS_STATUSES,
+  fulfillmentTypeMap,
 } from '../constants';
 import {
   buildUrl,
@@ -1008,7 +1009,7 @@ class RequestsRoute extends React.Component {
     const initialValues = dupRequest ||
     {
       requestType: 'Hold',
-      fulfilmentPreference: 'Hold Shelf',
+      fulfillmentPreference: fulfillmentTypeMap.HOLD_SHELF,
       createTitleLevelRequest: createTitleLevelRequestsByDefault,
     };
 
@@ -1188,7 +1189,7 @@ class RequestsRoute extends React.Component {
                 joinRequest: this.addRequestFields,
                 optionLists: {
                   addressTypes,
-                  fulfilmentTypes,
+                  fulfillmentTypes,
                   servicePoints,
                   cancellationReasons,
                 },